Login
Hey! So I am submitting to you what i did for the login using your framework... I got everything working and it is resposive aside form the username and password input which is looking werid. So i commented it out so you know where the problem is and can look at it.... let me know and thank you!